+++ sdl-core/src/main/scala/io/smartdatalake/util/crypt/EncryptDecryptGCM.scala
@@ -17,75 +17,12 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package io.smartdatalake.workflow.action.generic.transformer
-
-import org.apache.spark.sql.DataFrame
-import org.apache.spark.sql.expressions.UserDefinedFunction
-import org.apache.spark.sql.functions.{col, udf}
+package io.smartdatalake.util.crypt
 
 import java.security.SecureRandom
 import java.util.Base64
+import javax.crypto.spec.GCMParameterSpec
 import javax.crypto.{Cipher, SecretKey}
-import javax.crypto.spec.{GCMParameterSpec, SecretKeySpec}
-
-trait EncryptDecrypt extends Serializable {
-  protected def keyAsBytes: Array[Byte]
-
-  protected def encryptUDF: UserDefinedFunction = udf(encryptFunc _)
-
-  protected def decryptUDF: UserDefinedFunction = udf(decryptFunc _)
-
-  protected def encryptFunc(message: String): String
-
-  protected def decryptFunc(encryptedDataString: String): String
-
-  def encrypt(df: DataFrame, encryptColumns: Seq[String]): DataFrame = {
-    encryptColumns.foldLeft(df) {
-      case (dfTemp, colName) => dfTemp.withColumn(colName, encryptUDF(col(colName)))
-    }
-  }
-
-  def decrypt(df: DataFrame, encryptColumns: Seq[String]): DataFrame = {
-    encryptColumns.foldLeft(df) {
-      case (dfTemp, colName) => dfTemp.withColumn(colName, decryptUDF(col(colName)))
-    }
-  }
-
-  def generateAesKey(keyBytes: Array[Byte]): SecretKey = {
-    new SecretKeySpec(keyBytes, "AES")
-  }
-
-}
-
-class EncryptDecryptECB(keyBytes: Array[Byte]) extends EncryptDecrypt {
-  override protected val keyAsBytes: Array[Byte] = keyBytes
-  private val ALGORITHM_STRING: String = "AES/ECB/PKCS5Padding"
-
-  @transient lazy private val cipherEncrypt: Cipher = {
-    val cipher = Cipher.getInstance(ALGORITHM_STRING)
-    val secretKey: SecretKey = generateAesKey(keyAsBytes)
-    cipher.init(Cipher.ENCRYPT_MODE, secretKey)
-    cipher
-  }
-  @transient lazy private val cipherDecrypt: Cipher = {
-    val cipher = Cipher.getInstance(ALGORITHM_STRING)
-    val secretKey: SecretKey = generateAesKey(keyAsBytes)
-    cipher.init(Cipher.DECRYPT_MODE, secretKey)
-    cipher
-  }
-
-  override protected def encryptFunc(message: String): String = {
-    val data = cipherEncrypt.doFinal(message.getBytes())
-    Base64.getEncoder.encodeToString(data)
-  }
-
-  override protected def decryptFunc(encryptedDataString: String): String = {
-    val data = Base64.getDecoder.decode(encryptedDataString)
-    val message = cipherDecrypt.doFinal(data)
-    new String(message)
-  }
-
-}
 
 class EncryptDecryptGCM(keyBytes: Array[Byte]) extends EncryptDecrypt {
   override protected val keyAsBytes: Array[Byte] = keyBytes
@@ -97,7 +34,6 @@ class EncryptDecryptGCM(keyBytes: Array[Byte]) extends EncryptDecrypt {
   private val secureRandom = new SecureRandom()
 
   private val aesKey: SecretKey = generateAesKey(keyAsBytes)
-//  @transient lazy private val gcmParameterSpec = generateGcmParameterSpec()
 
   override protected def encryptFunc(message: String): String = {
     val gcmParameterSpec = generateGcmParameterSpec()
